feat(session): add option to restore sessions on server startup
