+++ .env.example
@@ -2,6 +2,7 @@
 # APPLICATION CONFIGURATION
 # ==================================
 PORT=3000
+RESTORE_SESSIONS_ON_START_UP=false
 
 # ==================================
 # DATABASE CONFIGURATION
