+++ src/api/controllers/instance.controller.js
@@ -2,6 +2,7 @@ const { WhatsAppInstance } = require('../class/instance')
 const fs = require('fs')
 const path = require('path')
 const config = require('../../config/config')
+const {Session} = require('../class/session')
 
 exports.init = async (req, res) => {
     const token = req.query.token
@@ -57,7 +58,7 @@ exports.qrbase64 = async (req, res) => {
 
 exports.info = async (req, res) => {
     const instance = WhatsAppInstances[req.query.key]
-    let data = ''
+    let data
     try {
         data = await instance.getInstanceDetail(req.query.key)
     } catch (error) {
@@ -66,25 +67,14 @@ exports.info = async (req, res) => {
     return res.json({
         error: false,
         message: 'Instance fetched successfully',
-        // instance_data: data,
         instance_data: data,
     })
 }
 
 exports.restore = async (req, res, next) => {
     try {
-        let restoredSessions = []
-        const instances = fs.readdirSync(path.join(__dirname, `../sessiondata`))
-        instances.map((file) => {
-            if (file.includes('.json')) {
-                restoredSessions.push(file.replace('.json', ''))
-            }
-        })
-        restoredSessions.map((key) => {
-            const instance = new WhatsAppInstance(key)
-            instance.init()
-            WhatsAppInstances[key] = instance
-        })
+        const session = new Session
+        let restoredSessions = await session.restoreSessions()
         return res.json({
             error: false,
             message: 'All instances restored',
